% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_runtime.R
\name{configure_gdal_home}
\alias{configure_gdal_home}
\title{Configure GDAL home for current session}
\usage{
configure_gdal_home(path, mode = c("option", "env"))
}
\arguments{
\item{path}{GDAL home directory path.}

\item{mode}{Either \code{"option"} or \code{"env"}.}
}
\value{
Invisibly returns the normalized GDAL home path.
}
\description{
Sets GDAL home for this session using either an R option or environment
variable. This does not write to user profile files.
}
