% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_setup.R
\name{gdal_uninstall}
\alias{gdal_uninstall}
\title{Uninstall managed GDAL resources}
\usage{
gdal_uninstall(
  lib = NULL,
  user_lib = FALSE,
  runtime = TRUE,
  package = TRUE,
  python = TRUE,
  force = FALSE
)
}
\arguments{
\item{lib}{Library containing the managed \code{gdalraster} package.}

\item{user_lib}{Use \code{.libPaths()[1]}.}

\item{runtime}{Remove the GDAL build runtime.}

\item{package}{Remove the managed \code{gdalraster} package.}

\item{python}{Remove managed embedded-Python \code{.pth} files.}

\item{force}{Confirm removal in non-interactive sessions.}
}
\value{
Invisibly, the removed paths.
}
\description{
Uninstall managed GDAL resources
}
