% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalraster.windows-package.R
\docType{package}
\name{gdalraster.windows-package}
\alias{gdalraster.windows}
\alias{gdalraster.windows-package}
\title{gdalraster.windows: Windows GDAL Runtime and gdalraster Bootstrap Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Installs and activates a self-contained GDAL runtime on Windows and provides helpers to install gdalraster from source against that GDAL instead of the default Rtools GDAL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/gdalraster.windows}
  \item \url{https://docs.jimbrig.com/gdalraster.windows/}
  \item Report bugs at \url{https://github.com/jimbrig/gdalraster.windows/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})
}

}
\keyword{internal}
